Included language files in the .jar file
Note: Resource folder still exists and is necessary as the configuration files are still in there. I do not plan to take them out and put them in the jar file, but at least make it so that the bot creates all directories himself if needed. If they already exist, he should be able to use them.
On the long term im planning to make it SQL based anyways so... Configloader Rework COMING!!!