First steps to MongoDB
Okay, I'm saying it how it is: Nothing is working at the moment. Maybe because I messed around with the most fundamental class of George: The Configloader. Well, at least we got MongoDB implemented now. Problem: I still do not know how everything works and how I am going to do it in the final release. But until then, I got a lot of time so I'll use that time and invest it into experimenting and researching lol. Have a great night!