Fixed the problem of configs not getting cached
MongoDB is now finally fully implemented, including a caching option to not completely ruin any internet connection. Further details can be looked up in the source code. Other than that: The poll command has been removed as of bad design. Furthermore I have included the font to render the levelcards in the levelcard resource folder so it shouldnt change when changing the computer or the operating system. And: I added a "UnhandledExceptionHandler" within the ConsoleEngine, so when not handled differently the ConsoleEngine should from now on spit all the errors out and keep the bot stable.