Implemented new language file structure
File structure in resources/languages now has to match up with the packet structure in the project.
Also I started working on implementing the Settings command. And don't worry about all the error in the actions package: I removed IOptionHolder as it doesnt make sense to ask for everything in the setting commands because of all the different explainations on what one specific option type could be used for. I therefore decided to ask for every option needed in the actions themselves. Therefore they are currently not compilable. But first the settings command with the most basic decisions, then all the actions.