Fixed bug from yesterday as example in SupportTalk.java
I've not gotten to fix the bug everywhere yet (Espacially because I'm pretty exhausted after also implementing the context menu operations), but I will so soon. I also planned to implement the context menu commands not only in said context menus but also as a slash command, just so the user may use the preferred way.