Combined operations to "StaticRoles" and "AutoMessages"
Fixed a massive bug inside ConfigVerifier.class as well, as I couldn't remove keys from a ConcurrentHashMap whilst .forEach'ing through it. In addition I implemented the {number} variable for the Join2Create channels, realized though, that a proper Join2Create rework is needed some time soon, as it's getting pretty messy and I don't want this to escalate when adding new variables etc. Maybe having one entitty class for Join2Create channels, instanciated on turning on the bot, would be enough, but we'll get to that in the future!