Started reworking ConsoleEngine.class
As mentioned in #45 our logging setup is VERY messy. So I started working on our own implementation of SLF4J today and removed the old logback dependency accordingly. The project itself is running again, but for some reason the ServiceProvider SLF4J need to recognize our implementation isn't found. I've just created a discussion on it over at the SLF4J repository, but we'll have to wait for an answer. Until then: The bot can be run safely, just without the detailed logging.