Permission framework rework & completion
Deleted the old permission framework again, as it was a duplicate of the one already implemented in JDA. The only exception would be the /configure command and all of it's configuration options, where I implemented a permission system.
Change to the main EventProcessor: If an insufficient permissions exception occurs it'll process it respectivly.
Change to Bot class: Adding all of the server util classes as event waiters had to be done different (my mistake)