Containing structure and first implementation
With this commit I am starting the work on the ConfigLoader Rework. This first step includes the basic structure setup of the rework as well as the first implementation of the GuildData class - A new class that is supposed to, with the UserData class and many else, process the JSONObjects coming from the MonogDB database to variables being contained inside those classes. It's simply just converting the JSONObject into Java Objects. The old system still works as I have not changed anything yet, but the new classes are in place and one of them even implemented (disregarding any tool methods though).