Finished implementing GuildData.class
This commit completes the GuildData.class for now. Meaning: All methods for data acquiring are implemented as well as the final compiling method which is run when saving the data object to the MongoDB database.
I might add compatibility for ids later, as it's not always handy to straight up use instances of TextChannels, VoiceChannels etc... This would probably nearly double the code though, so I'll have a look how much we actually would need it in the next couple of weeks. Think of it as an evaluation of work load against use frequency.