Started reworking the rework...
This results from a comment I made on #31 in which I explained the problem with saving objects such as Guilds, Roles, Users, Members and so on: They might get deleted, whilst the bot has saved them. In the future the bot will therefore save ids instead of those objects and convert these ids to objects as soon as they are requested from one of the data objects defined within the data classes.