Started work on implementing new structure
Bot.class is done, started with EventProcessor.class. Came up with the idea of outsourcing all the timeout checks within ModController.class to the EventProcessor and therefore changing it to an event based system. This does mean though, that an implementation for a startup check on the timeout status of all members is needed.