Removed DataContainer.interface and URL Leak
This commit contains the implementation of the secrets.json file, in which the URL to the auth server is stored. This is not safe, but keeps the URL from being puplicly available as it will only be send to developers or with the shipping .jar and the license key. This system will probably be replaced by some kind of encrypted, safer method to prevent the success of intentional leaking by costumers.
This commit also removes the remains of the old DataContainer.interface